requirements.txt
asteroid>=0.6.0
asteroid.egg==info
dcase_util>=0.2.19
h5py>=3.6.0
hydra-core>=1.3.1
jiwer>=2.5.1
librosa>=0.9.1
numexpr>=2.8.1
numpy>=1.21.5
omegaconf>=2.1.1
pandas>=1.4.2
pb_bss_eval>=0.0.2
pyannote.core>=4.4
pytorch_lightning>=1.4.9
scikit_optimize>=0.9.0
scipy>=1.8.0
setuptools>=58.0.4
SoundFile>=0.10.3.post1
speechbrain>=0.5.11
spy_der>=0.3.0
spyder>=5.4.0
torch>=1.11.0+cu113
torchaudio>=0.9.0
tqdm>=4.64.0
transformers>=4.25.1
